Rewrote amigaos_relocate_elfhandle() in amigaos-nat.c to do what it should do. Some other cleanups.
Also, added a new AmigaOS4 hook hack in elfread.c, in order to support relocated symbols for which no further debug information is available. Now, for each symbol, amigaos_get_symbol_addrs_hook() is called. The function is supposed to return the correct relocated address of the symbol in question.