Added new function attribute "lineartags" and pragma "amigaos tagtype".
Functions that have the lineartags attribute are assumed to be functions
with tags, in which case type checking is now enabled. That is, the value
that follow a specific tag are assumed to be a certain type. If this is
not the case, the compiler will warn.

The compiler can be taught about that types via the new pragma
#pragma amigaos tagtype(TYPE)
which is written before an "enum" or a "static const int" that defines
the number of the tag.

For instance:

#pragma amigaos tagtype(struct Screen *)
static const int WA_CustomScreen = WA_Dummy + 0x0D;

 or

#pragma amigaos tagtype(struct Screen *)
enum {WA_CustomScreen = WA_Dummy + 0x0D}; 

is possible now.

It is not very robust yet wrt to catching misuses of the new feature.
For instance, no check is done that the attribute is applied to varargs
functions.